Nicer Water
Now water behaves more like real water, it is like a liquid and sand falls through it.

+++ src/cells.rs
@@ -3,7 +3,7 @@ pub struct Cell {
     pub material: CellType,
     pub rgb: [u8; 3],
     pub selected: bool,
-    pub updated: u8,
+    pub updated: u16,
 }
 impl Default for Cell {
     fn default() -> Self {
